Refactor AiEntity, ClientCitation, SensitivityUsageInfo, TooManyStepsParameters
