+++ js/packages/teams-ai/src/types/index.ts
@@ -7,5 +7,9 @@
  */
 
 export * from './ActionCall';
+export * from './AIEntity';
+export * from './ClientCitation';
 export * from './CompletionConfig';
 export * from './PromptResponse';
+export * from './SensitivityUsageInfo';
+export * from './TooManyStepsParameters';
